+++ Modules/Microsoft365DSC/Examples/Resources/AADAuthenticationMethodPolicyAuthenticator/1-Create.ps1
@@ -24,17 +24,9 @@ Configuration Example
                     Id = 'Legal Team'
                     TargetType = 'group'
                 }
-                MSFT_AADAuthenticationMethodPolicyAuthenticatorExcludeTarget{
-                    Id = 'Paralegals'
-                    TargetType = 'group'
-                }
             );
             FeatureSettings       = MSFT_MicrosoftGraphmicrosoftAuthenticatorFeatureSettings{
                 DisplayLocationInformationRequiredState = MSFT_MicrosoftGraphAuthenticationMethodFeatureConfiguration{
-                    ExcludeTarget = MSFT_AADAuthenticationMethodPolicyAuthenticatorFeatureTarget{
-                        Id = 'all_users'
-                        TargetType = 'group'
-                    }
                     IncludeTarget = MSFT_AADAuthenticationMethodPolicyAuthenticatorFeatureTarget{
                         Id = 'all_users'
                         TargetType = 'group'
@@ -42,10 +34,6 @@ Configuration Example
                     State = 'default'
                 }
                 CompanionAppAllowedState = MSFT_MicrosoftGraphAuthenticationMethodFeatureConfiguration{
-                    ExcludeTarget = MSFT_AADAuthenticationMethodPolicyAuthenticatorFeatureTarget{
-                        Id = 'all_users'
-                        TargetType = 'group'
-                    }
                     IncludeTarget = MSFT_AADAuthenticationMethodPolicyAuthenticatorFeatureTarget{
                         Id = 'all_users'
                         TargetType = 'group'
@@ -53,10 +41,6 @@ Configuration Example
                     State = 'default'
                 }
                 DisplayAppInformationRequiredState = MSFT_MicrosoftGraphAuthenticationMethodFeatureConfiguration{
-                    ExcludeTarget = MSFT_AADAuthenticationMethodPolicyAuthenticatorFeatureTarget{
-                        Id = 'all_users'
-                        TargetType = 'group'
-                    }
                     IncludeTarget = MSFT_AADAuthenticationMethodPolicyAuthenticatorFeatureTarget{
                         Id = 'all_users'
                         TargetType = 'group'