+++ CHANGELOG.md
@@ -2,14 +2,19 @@
 
 # UNRELEASED
 
+
 * AADAdministrativeUnit
   * Fix the Update logic flow to get around a bug in Microsoft.Graph 2.11.1.
 * AADAuthenticationMethodPolicyX509
   * Added support for the  property for include targets
 * AADConditionalAccessPolicy
   * Added support for application filters in the conditions.
+* AADConditionalAccessPolicy
   * Implement Fix #3885. Manage Exclude Application.
     FIXES [[#3885](https://github.com/microsoft/Microsoft365DSC/issues/3885)]
+* AADGroupOwnerConsentSettings
+  * Initial release
+    Implements [#4112](https://github.com/microsoft/Microsoft365DSC/issues/4112)
 * EXOHostedContentFilterPolicy
   * Fix issue on parameters AllowedSenders, AllowedSenderDomains, BlockedSenders, 
     BlockSenderDomains if desired state is empty but current state is not empty.