+++ CHANGELOG.md
@@ -2,14 +2,12 @@
 
 # 1.24.110.1
 
-
 * AADAdministrativeUnit
   * Fix the Update logic flow to get around a bug in Microsoft.Graph 2.11.1.
 * AADAuthenticationMethodPolicyX509
   * Added support for the  property for include targets
 * AADConditionalAccessPolicy
   * Added support for application filters in the conditions.
-* AADConditionalAccessPolicy
   * Implement Fix #3885. Manage Exclude Application.
     FIXES [[#3885](https://github.com/microsoft/Microsoft365DSC/issues/3885)]
 * AADGroupOwnerConsentSettings
@@ -28,6 +26,8 @@
 * SCDLPComplianceRule
   * Fix type of AccessScope
     FIXES [#3463](https://github.com/microsoft/Microsoft365DSC/issues/3463)
+* TeamsTenantDialPlan
+  * FIXES [#3767](https://github.com/microsoft/Microsoft365DSC/issues/3767)
 
 # 1.24.103.1